+++ src/Stormancer.Plugins/GameSession/Stormancer.Gamesessions.Browser/GamesessionSearchState.cs
@@ -1,0 +1,7 @@
+﻿namespace Stormancer.Gamesessions.Browser
+{
+    internal class GamesessionSearchState
+    {
+        public object Document { get; internal set; }
+    }
+}