+++ src/Stormancer.Plugins/GameSession/Stormancer.Server.Plugins.GameSession/Changelog.rst
@@ -22,6 +22,7 @@ Added
 - Improved Game server system to enable sending custom arguments set during OnGameSessionStarting 
 - Enables complex game server image selection mechanisms through improved configuration and custom arguments.
 - Added IsDedicatedServer extension method to Session.
+- Added IGameSessionService.Scene property.
 
 Fixed
 *****